Consolidate client persistence validation
