+++ packages/shared-server/rallar-system/client-state/mutation/result-validation/validate-client-persistence.ts
@@ -1,4 +1,5 @@
 import { assertPersistableClientStateEvent } from '../../../state-events/postgres/client-state-event-row-codec.ts';
+import { validateClientEvent } from '../../client-state-contract-validation.ts';
 import { assertCanonicalClientStateIdempotencyRecord } from '../../persistence/client-state-repository-reads.ts';
 import {
     validatePersistedClientInstance,
@@ -6,13 +7,55 @@ import {
     validatePersistedClientSession
 } from '../../persistence/validate-persisted-client-state.ts';
 import { ClientMutationRejectedError } from '../../validation/client-mutation-rejection.ts';
+import {
+    decodeClientValidationRecord,
+    requireExactKeys,
+    type ClientValidationValue
+} from '../../validation/client-record-validation.ts';
+import { requireString } from '../../validation/client-string-validation.ts';
 import type {
     ClientMutationComputed,
     ClientMutationDomainWrite
 } from '../client-mutation-contracts.ts';
 import { computeClientPersistence } from '../compute/compute-client-persistence.ts';
 
-export function validateClientPersistence(
+export function validateClientPersistenceShape(value: ClientValidationValue): void {
+    const persistence = decodeClientValidationRecord(
+        value,
+        'Client mutation computed.persistence'
+    );
+    requireExactKeys(
+        persistence,
+        ['runtimeWrites', 'eventWrite'],
+        'Client mutation computed.persistence'
+    );
+    if (!Array.isArray(persistence.runtimeWrites)) {
+        throw new ClientMutationRejectedError(
+            'Client mutation computed.persistence.runtimeWrites must be an array'
+        );
+    }
+    persistence.runtimeWrites.forEach(validateClientRuntimeWriteShape);
+    if (persistence.eventWrite === null) {
+        return;
+    }
+    const eventWrite = decodeClientValidationRecord(
+        persistence.eventWrite,
+        'Client mutation computed.persistence.eventWrite'
+    );
+    requireExactKeys(
+        eventWrite,
+        ['event', 'workspaceKey', 'eventJson'],
+        'Client mutation computed.persistence.eventWrite'
+    );
+    validateClientEvent(eventWrite.event, 'Client mutation computed.persistence.eventWrite.event');
+    requireString(
+        eventWrite.workspaceKey,
+        'Client mutation computed.persistence.eventWrite.workspaceKey'
+    );
+    requireString(eventWrite.eventJson, 'Client mutation computed.persistence.eventWrite.eventJson');
+}
+
+export function validateExactClientPersistence(
     computed: Exclude<ClientMutationComputed, { outcome: 'idempotency-conflict'; }>
 ): void {
     if (
@@ -27,6 +70,30 @@ export function validateClientPersistence(
             JSON.stringify(computed.persistence)
     ) {
         throw new ClientMutationRejectedError('Client computed persistence differs');
+    }
+}
+
+function validateClientRuntimeWriteShape(value: ClientValidationValue, index: number): void {
+    const label = `Client mutation computed.persistence.runtimeWrites[${index}]`;
+    const write = decodeClientValidationRecord(value, label);
+    requireExactKeys(
+        write,
+        ['kind', 'namespace', 'key', 'value', 'expireAtIsoTimestamp', 'expectedRevision'],
+        label
+    );
+    requireString(write.namespace, `${label}.namespace`);
+    requireString(write.key, `${label}.key`);
+    requireString(write.value, `${label}.value`);
+    requireString(write.expireAtIsoTimestamp, `${label}.expireAtIsoTimestamp`);
+    if (write.kind === 'insert' && write.expectedRevision === null) {
+        return;
+    }
+    if (
+        write.kind !== 'update' ||
+        !Number.isSafeInteger(write.expectedRevision) ||
+        (write.expectedRevision as number) < 0
+    ) {
+        throw new ClientMutationRejectedError(`${label} guard is invalid`);
     }
 }
